+++ src/test/html/html-script-scanner_test.ts
@@ -13,12 +13,18 @@
  */
 
 import {assert} from 'chai';
+import * as path from 'path';
 
+import {Analyzer} from '../../core/analyzer';
 import {HtmlVisitor} from '../../html/html-document';
 import {HtmlParser} from '../../html/html-parser';
 import {HtmlScriptScanner} from '../../html/html-script-scanner';
+import {JavaScriptDocument} from '../../javascript/javascript-document';
+import {Analysis} from '../../model/analysis';
 import {ScannedImport, ScannedInlineDocument} from '../../model/model';
+import {FSUrlLoader} from '../../url-loader/fs-url-loader';
 
+const fixturesDir = path.resolve(__dirname, '../static');
 suite('HtmlScriptScanner', () => {
 
   suite('scan()', () => {
@@ -39,11 +45,11 @@ suite('HtmlScriptScanner', () => {
       const {features} = await scanner.scan(document, visit);
       assert.equal(features.length, 2);
       assert.instanceOf(features[0], ScannedImport);
-      const feature0 = <ScannedImport>features[0];
+      const feature0 = features[0] as ScannedImport;
       assert.equal(feature0.type, 'html-script');
       assert.equal(feature0.url, 'foo.js');
       assert.instanceOf(features[1], ScannedInlineDocument);
-      const feature1 = <ScannedInlineDocument>features[1];
+      const feature1 = features[1] as ScannedInlineDocument;
       assert.equal(feature1.type, 'js');
       assert.equal(feature1.contents, `console.log('hi')`);
       assert.deepEqual(feature1.locationOffset, {line: 2, col: 18});
@@ -59,9 +65,60 @@ suite('HtmlScriptScanner', () => {
       const {features} = await scanner.scan(document, visit);
       assert.equal(features.length, 1);
       assert.instanceOf(features[0], ScannedImport);
-      const feature0 = <ScannedImport>features[0];
+      const feature0 = features[0] as ScannedImport;
       assert.equal(feature0.type, 'html-script');
       assert.equal(feature0.url, '/aybabtu/foo.js');
+    });
+
+    suite('modules', () => {
+      const urlLoader = new FSUrlLoader(fixturesDir);
+      const analyzer = new Analyzer({urlLoader});
+      let analysis: Analysis;
+
+      before(async() => {
+        analysis = await analyzer.analyze(['js-modules.html']);
+      });
+
+      test('finds external module scripts', () => {
+        const htmlScripts = [...analysis.getFeatures({kind: 'html-script'})];
+        assert.equal(htmlScripts.length, 1);
+        const js = htmlScripts[0].document.parsedDocument as JavaScriptDocument;
+        assert.equal(js.url, 'javascript/module.js');
+        assert.equal(js.parsedAsSourceType, 'module');
+        assert.equal(
+            js.contents.trim(), `import * as submodule from './submodule.js';`);
+      });
+
+      test('finds inline module scripts', () => {
+        const inlineDocuments =
+            [...analysis.getFeatures({kind: 'inline-document'})];
+        assert.equal(inlineDocuments.length, 1);
+        const js = inlineDocuments[0].parsedDocument as JavaScriptDocument;
+        assert.equal(js.url, 'js-modules.html');
+        assert.equal(js.parsedAsSourceType, 'module');
+        assert.equal(
+            js.contents.trim(),
+            `import * as something from './javascript/module-with-export.js';`);
+      });
+
+      test('follows import statements in modules', async() => {
+        const jsImports = [...analysis.getFeatures({kind: 'js-import'})];
+        assert.equal(jsImports.length, 2);
+
+        // import statement in inline module script in 'js-modules.html'
+        const js0 = jsImports[0].document.parsedDocument as JavaScriptDocument;
+        assert.equal(js0.url, 'javascript/module-with-export.js');
+        assert.equal(js0.parsedAsSourceType, 'module');
+        assert.equal(
+            js0.contents.trim(), `export const someValue = 'value goes here';`);
+
+        // import statement in external module script 'javascript/module.js'
+        const js1 = jsImports[1].document.parsedDocument as JavaScriptDocument;
+        assert.equal(js1.url, 'javascript/submodule.js');
+        assert.equal(js1.parsedAsSourceType, 'module');
+        assert.equal(
+            js1.contents.trim(), `export const subThing = 'sub-thing';`);
+      });
     });
   });
 });